fifth commit

+++ database/migrations/2024_09_01_130830_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('company_id');
+            $table->unsignedBigInteger('company_id');
             $table->string('product_name');
             $table->integer('price');
             $table->integer('stock');
